pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prodogotchi"
version = "0.1.0"
dependencies = ["pyyaml"]

[project.scripts]
prodogotchi = "prodogotchi.cli:main"
